requirements.txt
flask
eciespy
qrcode
opencv-python-headless
cryptography
pillow
requests
numpy
gunicorn
